@@ -1,10 +1,10 @@
 name: emma_flutter_sdk
 description: EMMA SDK implementation for flutter. This package contains native bindings for our official SDK's. You can find more info at https://support.emma.io or https://www.emma.io
-version: 1.2.0
+version: 1.3.0
 homepage: https://www.emma.io
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0"
 
 dependencies:
